Use the actually documented defaults in the cloudscale_server resource
If using use_private_network = False or use_ipv6 = False the default was still
always true, because we checked it with ResourceData.GetOk, which only returns
ok if the value is non-zero. One possible fix in the future would have been to
use ResourceData.GetOkExists instead. However this method only exists on the
master branch and not in any current terraform release. Therefore we are just
using defaults and ResourceData.Get.

Added a test to avoid the issue in the future.