Use the actually documented defaults in the cloudscale_server resource (#1)
* Use the actually documented defaults in the cloudscale_server resource

If using use_private_network = False or use_ipv6 = False the default was still
always true, because we checked it with ResourceData.GetOk, which only returns
ok if the value is non-zero. One possible fix in the future would have been to
use ResourceData.GetOkExists instead. However this method only exists on the
master branch and not in any current terraform release. Therefore we are just
using defaults and ResourceData.Get.

Added a test to avoid the issue in the future.

* Make the PrivateNetwork test clearer and check everything we want to check.

* Use lists instead of sets to represent cloudscale data

An issue was discovered where only the first interface of two was listed. I'm
not sure how TypeSets in terraform do uniqueness, but they do it definitely
wrong in this case. Lists make sense anyway, because the cloudscale.ch
provider ensures uniqueness anyway. An advantage is also that the cloudscale.ch
API sorts lists already.

There's a test now that ensures that private and public interfaces appear both
in the terraform output.

* Use TypeList instead of TypeSet for computed attributes

Related to the previous change with the TypeSet issue, we avoid sets here as well.

* Use GetOkExists instead of GetOk for boolean values

Since terraform was updated to a very recent commit we can now use it.